Fix SSO auth not redirecting to /authorize - When an OAuth app requests a login and the user wants to login via SSO it will now properly redirect to the authorize URL - Refactor the SSO authenticators a bit, so more code is shared
